requirement.txt
accelerate==0.29.2
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
biopython==1.79
bleach==6.1.0
blessed==1.20.0
bokeh==3.3.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.0
comm==0.1.3
contourpy==1.0.7
croniter==1.4.1
cycler==0.11.0
dask==2023.6.1
dask_labextension==7.0.0
dateutils==0.6.12
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.1
defusedxml==0.7.1
dill==0.3.7
distributed==2023.6.1
einops==0.7.0
executing==1.2.0
fair-esm==2.0.0
fastapi==0.101.1
fastjsonschema==2.18.1
filelock==3.10.7
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.5.0
h11==0.14.0
huggingface-hub==0.22.2
idna==3.4
importlib-metadata==6.6.0
inquirer==3.1.3
ipykernel==6.22.0
ipython==8.11.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.9.1
jupyter_server_proxy==4.1.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kiwisolver==1.4.4
lightning==2.0.7
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lit==16.0.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mdurl==0.1.2
mistune==3.0.2
modin==0.22.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.0
notebook_shim==0.2.3
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
ordered-set==4.1.0
overrides==7.4.0
packaging==23.0
pandarallel==1.6.5
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==3.2.0
prometheus-client==0.18.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.14.0
PyJWT==2.8.0
pyparsing==3.0.9
pyreadstat==1.2.3
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
readchar==4.0.5
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.12.0
safetensors==0.4.3
SciencePlots==2.1.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.1
Send2Trash==1.8.2
sentencepiece==0.2.0
setuptools==67.6.0
simpervisor==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
swifter==1.4.0
sympy==1.11.1
tblib==1.7.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchmetrics==1.0.3
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.39.3
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
UpSetPlot==0.9.0
uri-template==1.3.0
urllib3==1.26.15
uvicorn==0.23.2
venn==0.1.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
wheel==0.38.4
xyzservices==2023.10.1
yarl==1.9.2
zict==3.0.0
zipp==3.15.0
